Modularize alerts into useAlert custom hook
